Divide map class.

+++ Base/base.hpp
@@ -7,9 +7,9 @@ void LogError(const std::string& msg);
 void LogSdlError(const std::string& msg);
 void LogTtfError(const std::string& msg);
 
-const float EPS = 0.0001;
+const float EPS = 0.0001f;
 
-const float TO_DEG = 180.0f / M_PI;
-const float TO_RAD = M_PI / 180.0f;
+const float TO_DEG = 180.0f / static_cast<float>(M_PI);
+const float TO_RAD = static_cast<float>(M_PI) / 180.0f;
 
 #endif // BASE_H
